Update remote host updating process in Jenkinsfile
This commit alters the process of updating the remote host in the Jenkinsfile, now utilizing a secure login. Moreover, it also introduces a new logback configuration for better logging, and a version increment in pom.xml. Additionally, the Dockerfile has been updated to include the copying of pom.xml file.